fastapi_ia_app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.110.0
filelock==3.18.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.78.1
pydantic==2.6.4
pydantic_core==2.16.3
pyee==13.0.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.36.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
